requirements.txt
asgiref==3.8.1
Django==4.2.11
djangorestframework==3.15.1
sqlparse==0.4.4
typing-extensions==4.10.0

pip==20.3.4
keyring==23.0.1
colorama==0.4.4
PyYAML==5.3.1
Jinja2==2.11.3
requests==2.31.0
setuptools==52.0.0
pytz==2021.1